requirements.txt
anyio==3.6.2
certifi==2022.9.24
colorlog==6.7.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.2
httpx==0.23.1
hyperframe==6.0.1
idna==3.4
rfc3986==1.5.0
sniffio==1.3.0
tabulate==0.9.0
progressbar2==4.2.0

